MolTrack: Registration enhancements (#56)
* MolTrack: Refactor registrars to prevent automatic updates of compound details

* MolTrack: Add support for returning a limited number of search results

* MolTrack: Use direct db field for structure instead of reserved field name

* MolTrack: Update the validator for operands

* MolTrack: Add limit to every search endpoint

* MolTrack: Support updating property friendly names

* MolTrack: Add semantic type Molecule

* MolTrack: Update compound-related endpoints to perform actions using corporate_compound_id

* MolTrack: Update batch-related endpoints to perform actions using corporate_batch_id

* MolTrack: Enhance compound endpoints to support retrieval, update, and deletion by any synonym

* MolTrack: Enhance batch endpoints to support retrieval and deletion by any synonym

* MolTrack: Update tests to align with new endpoint implementation

* MolTrack: Allow setting semantic type name when creating a property

* MolTrack: Restrict update and delete operations to corporate ids only

* MolTrack: Add endpoint for automatic column mapping

* MolTrack: Return semantic types for direct fields like dynamic fields

* MolTrack: Resolve issues introduced after merge

* MolTrack: Updates after merge

* MolTrack: Fix invalid validator test

+++ app/crud/batches.py
@@ -3,29 +3,35 @@ from sqlalchemy import and_
 from sqlalchemy.orm import Session
 from sqlalchemy.orm import joinedload
 from app.crud.properties import enrich_properties
-from app import models
+from app import crud, models
 
 
 def enrich_batch(batch: models.Batch) -> models.BatchResponse:
     return models.BatchResponse(**batch.dict(), properties=enrich_properties(batch, "batch_details", "batch_id"))
 
 
-def get_batch(db: Session, corporate_batch_id: str):
+def get_batch_by_synonym(db: Session, property_value: str, property_name: str = None, enrich: bool = True):
+    if not property_value:
+        return None
+
+    filters = [models.Property.semantic_type_id == 1, models.BatchDetail.value_string == property_value]
+
+    if property_name:
+        filters.append(models.Property.name == property_name)
+
     batch = (
         db.query(models.Batch)
         .join(models.Batch.batch_details)
         .join(models.BatchDetail.property)
         .options(joinedload(models.Batch.batch_details).joinedload(models.BatchDetail.property))
-        .filter(
-            and_(models.Property.name == "corporate_batch_id", models.BatchDetail.value_string == corporate_batch_id)
-        )
+        .filter(and_(*filters))
         .first()
     )
 
     if not batch:
         return None
 
-    return enrich_batch(batch)
+    return enrich_batch(batch) if enrich else batch
 
 
 def get_batches(db: Session, skip: int = 0, limit: int = 100):
@@ -37,13 +43,19 @@ def get_batches_by_compound(db: Session, compound_id: int, skip: int = 0, limit:
     return db.query(models.Batch).filter(models.Batch.compound_id == compound_id).offset(skip).limit(limit).all()
 
 
-def delete_batch(db: Session, batch_id: int):
-    db_batch = db.get(models.Batch, batch_id)
+def delete_batch_by_synonym(
+    db: Session,
+    property_value: str,
+    property_name: str,
+):
+    db_batch = get_batch_by_synonym(db, property_value, property_name, False)
     if db_batch is None:
         raise HTTPException(status_code=404, detail="Batch not found")
 
-    db.query(models.BatchDetail).filter(models.BatchDetail.batch_id == batch_id).delete(synchronize_session=False)
-    db.query(models.BatchAddition).filter(models.BatchAddition.batch_id == batch_id).delete(synchronize_session=False)
+    assay_results = crud.get_all_assay_results_for_batch(db, db_batch.id)
+    if assay_results:
+        raise HTTPException(status_code=400, detail="Batch has dependent assay results")
+
     db.delete(db_batch)
     db.commit()
     return db_batch
